# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateRotatingSavingsGroupRequest:
    properties:
      account_id:
        type: string
      contribution_amount:
        type: number
      cycle_frequency:
        description: weekly, monthly
        type: string
      fixed_interest_amount:
        type: number
      member_count:
        type: integer
      name:
        type: string
      start_date:
        description: YYYY-MM-DD
        type: string
      status:
        type: string
      user_slots:
        type: integer
    type: object
  dto.CreateSavingsRequest:
    properties:
      auto_renew:
        type: boolean
      interest_rate:
        type: string
      maturity_date:
        type: string
      name:
        type: string
      parent_account_id:
        type: string
      principal:
        type: string
      savings_account_id:
        type: string
      start_date:
        type: string
      term_months:
        type: integer
    type: object
...
  dto.PatchSavingsRequest:
    properties:
      auto_renew:
        type: boolean
      interest_rate:
        type: string
      maturity_date:
        type: string
      name:
        type: string
      principal:
        type: string
      savings_account_id:
        type: string
      status:
        type: string
      term_months:
        type: integer
    type: object
...
  dto.RotatingSavingsContributionRequest:
    properties:
      account_id:
        type: string
      amount:
        type: string
      collected_fee_per_slot:
        type: number
      cycle_no:
        type: integer
      due_date:
        type: string
      kind:
        description: contribution, payout, collected
        type: string
      note:
        type: string
      occurred_date:
        type: string
      occurred_time:
        type: string
      slots_taken:
        type: integer
    type: object
...
  dto.UpdateRotatingSavingsGroupRequest:
    properties:
      account_id:
        type: string
      contribution_amount:
        type: number
      fixed_interest_amount:
        type: number
      name:
        type: string
      payout_cycle_no:
        type: integer
      status:
        type: string
    type: object
...
  entity.Savings:
    properties:
      accrued_interest:
        type: string
      auto_renew:
        type: boolean
      closed_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      interest_rate:
        type: string
      maturity_date:
        type: string
      parent_account_id:
        type: string
      principal:
        type: string
      savings_account_id:
        type: string
      start_date:
        type: string
      status:
        type: string
      term_months:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.ErrorBody:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  response.ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/response.ErrorBody'
    type: object
host: api.goen.wisteria.local
info:
  contact: {}
  description: REST API for Goen Project
  title: Goen API
  version: "1.0"
paths:
...
  /rotating-savings:
    get:
      description: Retrieve a list of Rotating Savings (Hụi) groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Rotating Savings Groups
      tags:
      - RotatingSavings
    post:
      consumes:
      - application/json
      description: Define a new rotating savings (Hụi) group
      parameters:
      - description: Rotating Group Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRotatingSavingsGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create Rotating Savings Group
      tags:
      - RotatingSavings
  /rotating-savings/{id}:
    delete:
      description: Clean up a specific rotating group structure
      parameters:
      - description: Rotating Savings Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete Rotating Savings Group
      tags:
      - RotatingSavings
    get:
      description: Retrieve detailed aggregation of a Rotating Savings group
      parameters:
      - description: Rotating Savings Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get Rotating Savings Group
      tags:
      - RotatingSavings
    patch:
      consumes:
      - application/json
      description: Partially update metadata properties for a group
      parameters:
      - description: Rotating Savings Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Group Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRotatingSavingsGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update Rotating Savings Group
      tags:
      - RotatingSavings
  /rotating-savings/{id}/contributions:
    post:
      consumes:
      - application/json
      description: Make a contribution (bidding/paying) within a group cycle
      parameters:
      - description: Rotating Savings Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Contribution Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RotatingSavingsContributionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create Contribution
      tags:
      - RotatingSavings
  /rotating-savings/{id}/contributions/{contributionId}:
    delete:
      description: Void a rotating savings contribution history by ID
      parameters:
      - description: Rotating Savings Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Contribution ID
        in: path
        name: contributionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete Contribution
      tags:
      - RotatingSavings
  /savings:
    get:
      description: Retrieve personal savings goals/accounts for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Savings'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List Savings
      tags:
      - Savings
    post:
      consumes:
      - application/json
      description: Create a personalized savings goal or instrument
      parameters:
      - description: Savings Creation Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSavingsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Savings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create Savings
      tags:
      - Savings
  /savings/{id}:
    delete:
      description: Delete a savings goal by ID
      parameters:
      - description: Savings ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete Savings
      tags:
      - Savings
    get:
      description: Retrieve details for a specific savings instance
      parameters:
      - description: Savings ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Savings'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get Savings
      tags:
      - Savings
    patch:
      consumes:
      - application/json
      description: Partially update specific savings settings
      parameters:
      - description: Savings ID
        in: path
        name: id
        required: true
        type: string
      - description: Savings Update Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PatchSavingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Savings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update Savings
      tags:
      - Savings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"